Move convert function in different module
